Rename profile.factory to profile.helpers
