+++ perun/logic/runner.py
@@ -9,7 +9,7 @@ import perun.logic.pcs as pcs
 import perun.logic.config as config
 import perun.logic.commands as commands
 import perun.logic.index as index
-import perun.profile.factory as profile
+import perun.profile.helpers as profile
 import perun.utils as utils
 import perun.utils.log as log
 import perun.utils.decorators as decorators
