+++ perun/view/flow/bokeh_factory.py
@@ -6,7 +6,7 @@ with demandimport.enabled():
     import bokeh.models as models
     import pandas
 
-import perun.profile.factory as profiles
+import perun.profile.helpers as profiles
 
 import perun.profile.convert as convert
 import perun.utils.bokeh_helpers as bokeh_helpers
